refactor: separate Post from webtoon
